+++ libraries/Bluefruit52Lib/src/BLEScanner.h
@@ -66,10 +66,10 @@ public:
   void setStopCallback(stop_callback_t fp);
 
   /*------------- Data Parser -------------*/
-  uint8_t* parseReportByType(uint8_t const* scandata, uint8_t scanlen, uint8_t type, uint8_t* result_len);
-  uint8_t* parseReportByType(const ble_gap_evt_adv_report_t* report, uint8_t type, uint8_t* result_len);
+  uint8_t parseReportByType(const uint8_t* scandata, uint8_t scanlen, uint8_t type, uint8_t* buf, uint8_t bufsize = 0);
+  uint8_t parseReportByType(const ble_gap_evt_adv_report_t* report, uint8_t type, uint8_t* buf, uint8_t bufsize = 0);
 
-  bool    checkReportForUuid(const ble_gap_evt_adv_report_t* adv_report, BLEUuid ble_uuid);
+  bool    checkReportForUuid(const ble_gap_evt_adv_report_t* report, BLEUuid ble_uuid);
 
   /*------------------------------------------------------------------*/
   /* INTERNAL USAGE ONLY